+++ components/molecules/HighlightInput/highlight-input-form.tsx
@@ -15,6 +15,7 @@ import { BiGitMerge } from "react-icons/bi";
 import { VscIssues } from "react-icons/vsc";
 import { A11y, Pagination } from "swiper/modules";
 import Skeleton from "react-loading-skeleton";
+import clsx from "clsx";
 import Button from "components/shared/Button/button";
 import Tooltip from "components/atoms/Tooltip/tooltip";
 
@@ -564,7 +565,7 @@ const HighlightInputForm = ({ refreshCallback }: HighlightInputFormProps): JSX.E
                 </span>
               )}
             </p>
-            <div className="flex flex-col gap-2 p-2 text-sm bg-white border rounded-lg">
+            <div className="flex flex-col gap-2 p-2 text-sm bg-white border rounded-lg focus-within:outline-none focus-within:ring focus-within:border-orange-500 focus-within:ring-orange-100">
               <TypeWriterTextArea
                 className={`resize-y min-h-[80px] max-h-99 font-normal placeholder:text-slate-400 text-light-slate-12 placeholder:font-normal placeholder:text-sm transition focus:outline-none rounded-lg ${
                   !isDivFocused ? "hidden" : ""
@@ -642,7 +643,8 @@ const HighlightInputForm = ({ refreshCallback }: HighlightInputFormProps): JSX.E
                 </Tooltip>
                 <TextInput
                   id="highlight-link-input"
-                  className={`text-sm shadow-none h-10 flex-none ${!isHighlightURLValid && "border-red-500"}`}
+                  state={isHighlightURLValid ? "valid" : "invalid"}
+                  className={clsx("text-sm shadow-none h-10 flex-none", errorMsg ? "border-red-500" : "")}
                   value={highlightLink}
                   handleChange={(value) => {
                     setHighlightLink(value);